+++ docs/history.rst
@@ -89,4 +89,14 @@ History
 * hf.data_catalog(site) retrieves history of data collected at site.
 * Dropped Python 3.4 & 3.5 support, added 3.8.
 
+0.2.1 (2021-05-28)
+------------------
 
+* Moved from TravisCI to Github Actions.
+* Added Python 3.9 support.
+* Created a verbose mode that is on by default.
+* Added the flood stage service from the USGS waterwatch (iem)
+* save & read NWIS data in json.gzip files
+* Added ability to read multiple instrument methods at the same times
+* bugfix: divide by zero error when you request data from many sites
+* bugfix: save_parquet